Moved parsers to parsers package
