+++ src/main/java/ru/zhelper/zhelper/services/parser/ZakupkiParser44Fz.java
@@ -1,4 +1,4 @@
-package ru.zhelper.zhelper.services;
+package ru.zhelper.zhelper.services.parser;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
